requirements.txt
# Core Scientific Computing
numpy==1.26.4
pandas==2.2.0
scipy==1.12.0

# Geospatial Processing
rasterio==1.3.9
geopandas==0.14.3
shapely==2.0.3
pyproj==3.6.1
fiona==1.9.5

# Satellite Data Processing
netCDF4==1.6.5
h5py==3.10.0
gdal==3.8.4

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3

# Visualization
matplotlib==3.8.2
seaborn==0.13.2
folium==0.15.1
plotly==5.18.0

# KML/XML Processing
lxml==5.1.0
simplekml==1.3.6

# Utilities
pyyaml==6.0.1
tqdm==4.66.1
pillow==10.2.0

# Optional: GPU acceleration for ML (if CUDA available)
# Uncomment if you want to use GPU for model training
# torch==2.2.0
# torchvision==0.17.0
